% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19_at.R
\name{get_first_date}
\alias{get_first_date}
\title{Get date of first available data point in db}
\usage{
get_first_date(wikipedia_table)
}
\arguments{
\item{wikipedia_table}{Data to be checked}
}
\value{
The date of the first data point
}
\description{
Get date of first available data point in db
}
\examples{
get_first_date(scrape_wikipedia_at()[[1]])
}
