% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19_at.R
\name{get_last_date}
\alias{get_last_date}
\title{Get date of last available data point in db}
\usage{
get_last_date(wikipedia_table)
}
\arguments{
\item{wikipedia_table}{Data to be checked}
}
\value{
The date of the last data point
}
\description{
Get date of last available data point in db
}
\examples{
get_last_date(scrape_wikipedia_at()[[1]])
}
