% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19_at.R
\name{plot_compare_at_it}
\alias{plot_compare_at_it}
\title{Compares Austrian and Italian infection data}
\usage{
plot_compare_at_it(days_shift = 8)
}
\arguments{
\item{days_shift}{How many days the Austrian data is shifted}
}
\value{

}
\description{
The function creates a plot comparing Austrian and Italian infection data.
It saves the plot to COMPARISON_AT_IT_FILENAME
}
\examples{
plot_compare_at_it()
}
