% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19_at.R
\name{plot_infected_tests_ratio}
\alias{plot_infected_tests_ratio}
\title{Plots timeseries of the infected:number of tests ratio.}
\usage{
plot_infected_tests_ratio(wikipedia_table)
}
\arguments{
\item{wikipedia_table}{Austrian infection data to be used}
}
\value{

}
\description{
Plots figures and saves it to INFECTED_TEST_RATIO_FILENAME
}
\examples{
plot_infected_tests_ratio(scrape_wikipedia_at()[[1]])
}
