% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19_at.R
\name{plot_number_tests}
\alias{plot_number_tests}
\title{Plots timeseries of number of tests}
\usage{
plot_number_tests(wikipedia_table)
}
\arguments{
\item{wikipedia_table}{Austrian infection data to be used}
}
\value{

}
\description{
Plots figures and saves it to NMB_TESTS_FILENAME
}
\examples{
plot_number_tests(scrape_wikipedia_at()[[1]])
}
