% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19_at.R
\name{plot_overview_at}
\alias{plot_overview_at}
\title{Plots an overview of Austrian infection data}
\usage{
plot_overview_at(wikipedia_table)
}
\arguments{
\item{wikipedia_table}{Austrian infection data to be used}
}
\value{

}
\description{
Plots figures and saves it to OVERVIEW_FILENAME
}
\examples{
plot_overiew_at(scrape_wikipedia_at()[[1]])
}
