% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19_at.R
\name{plot_prediction}
\alias{plot_prediction}
\title{Creates a plot showing a model fit to the data. The model can be chosen arbitrarily}
\usage{
plot_prediction(
  wikipedia_table,
  final_date,
  steps = 3,
  transformation_f = log,
  transformation_f_inverse = exp
)
}
\arguments{
\item{wikipedia_table}{Data to be fitted. Has to contain columns Datum and Infektionen}

\item{final_date}{Up to which date should be predicted (POSIXct date)}

\item{steps}{How many models should be fitted by reducing the amount of data points.
I.e. 3 indicates that one model with all data points, one model with all
data points minus the last one and one model with all data points minus
the two last ones is fitted.}

\item{transformation_f}{Function that transforms input data}

\item{transformation_f_inverse}{Inverse of trnasformation_f. This function indicates the kind of model. I.e. if using exp here, we estimate an exponential model.}
}
\value{

}
\description{
The function fits a generic linear model between Datum and Infektionen
Before fitting, Infektionen are transformed according to the given transformation functions.
The model is fitted for different amounts of data points.
The models are plotted together with data points and the resulting plot is saved to MODEL_FILENAME
}
