% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19_at.R
\name{scrape_wikipedia_at}
\alias{scrape_wikipedia_at}
\title{Download, clean and save data on infections in Austria from wikipedia}
\usage{
scrape_wikipedia_at(wikipedia_url = WIKIPEDIA_URL_AT)
}
\arguments{
\item{url}{Url to the wikipedia page. Is preconfigured and may only be changed if wikipedia URL changed.}
}
\value{
A list of two tibbles: the first elment contains a tibble with up-to-date data, the second element contains a tibble with the last data set loaded from disk.
                               Both tibbles have the same format and include all information from the wikipedia table.
}
\description{
Downloads the data and saves a feather file with the data in data/data.feather
}
\examples{
scrape_wikipedia_at()
}
