% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19_at.R
\name{scrape_wikipedia_it}
\alias{scrape_wikipedia_it}
\title{Download, clean and save data on infections in Italy from wikipedia}
\usage{
scrape_wikipedia_it(wikipedia_url = WIKIPEDIA_URL_IT)
}
\arguments{
\item{url}{Url to the wikipedia page. Is preconfigured and may only be changed if wikipedia URL changed.}
}
\value{
A tibble with infection data for Italy
}
\description{
Download, clean and save data on infections in Italy from wikipedia
}
\examples{
scrape_wikipedia_it()
}
